requirements.txt
fastapi==0.115.12
passlib==1.7.4
pydantic==2.10.6
PyJWT==2.10.1
PyYAML==6.0.2
SQLAlchemy==2.0.39
uvicorn==0.34.0
